% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_450k}
\alias{coef_450k}
\title{Coefficients of 450K model}
\format{
A list of 3 vectors, each contains 123 coefficients
}
\usage{
data("coef_450k")
}
\description{
Coefficents of EpiSmokEr 450K model from Bollepalli et al., 2019.
}
\details{
\itemize{
  \item coef.450k.ls: Coefficients of intercept, sex term, and 121 CpGs for current smoker (cs), former smoker (fs), and never smoker (ns).
}
}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\keyword{coefficients}
