% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coef_EPIC}
\alias{coef_EPIC}
\title{Coefficients of EPIC model}
\format{
A list of 3 vectors, each contains 513 coefficients
}
\usage{
data("coef_EPIC")
}
\description{
Coefficents of EpiSmokEr2 EPIC model
}
\details{
\itemize{
  \item coef.epic.ls: Coefficients of intercept, sex term, and 511 CpGs for current smoker (cs), former smoker (fs), and never smoker (ns).
}
}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\keyword{coefficients}
