% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Reads IDAT files}
\usage{
loadData(idatPath)
}
\arguments{
\item{idatPath}{Requires a path to the folder with the IDAT files and sample sheet included.}
}
\value{
A RGChannelSet object
}
\description{
loadData function reads IDAT files and saves as a minfi RGSet object. Details can be found in the help page of *read.metharray.sheet* and *read.metharray.exp* function from *minfi* R package.
}
\examples{
#RGset <- loadData(idatPath = idatPath)


}
