% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeData.R
\name{normalizeData}
\alias{normalizeData}
\title{Color and channel specific quantile normalization}
\usage{
normalizeData(RGset, normMethod=c("QN","ILM","SQN","ALL"), array = c("450K","EPIC"))
}
\arguments{
\item{RGset}{RGChannelSet object from minfi.}

\item{normMethod}{The method for quantile normalization, can be chosen from "QN", "ILM", "SQN", "ALL". "QN" is quantile normalization for each color channel and probe type with training data as reference. "ILM" and "SQN" are from minfi. "ALL" means using all 3 methods.}

\item{array}{Array type for the input CpGs. "450K" for Illumina HumanMethylation450 data and "EPIC" for Illumina MethylationEPIC data.}
}
\value{
\item{dataset_QN}{Beta value matrix after quantile normalization. Available if normMethod = "QN" or "ALL".}

\item{dataset_ILM}{Beta value matrix after ILM normalization. Available if normMethod = "ILM" or "ALL".}

\item{dataset_SQN}{Beta value matrix after SQN normalization. Available if normMethod = "SQN" or "ALL".}
}
\description{
Performs quantile normaliation on signal intensity values for each color channel and probe type for Illumina 450K or EPIC data.
}
\details{
This function does quantile normalization to input data.
}
\examples{
#dataset <- normalizeData(RGset, normMethod="QN", array = "450k")

}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\author{
Sailalitha Bollepalli, Tianyu Zhu
}
