% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_epic}
\alias{probe_epic}
\title{Probe list for EPIC data}
\format{
A list of 3 vectors
}
\usage{
data("probe_epic")
}
\description{
Type I (two color channel) and Type II probe names of EPIC training data (YFS)
}
\details{
\itemize{
  \item probe.epic.ls: Type I (two color channel) and Type II probe names of EPIC training data (YFS)
}
}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\keyword{probes}
