% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quan_450k}
\alias{quan_450k}
\title{Quantiles of 450K data}
\format{
A list of 6 vectors
}
\usage{
data("quan_450k")
}
\description{
Quantiles of intensity values based on each probe type and color channel from training data (DILGOM)
}
\details{
\itemize{
  \item quantile.450k.ls: Quantiles of intensity values based on each probe type and color channel
}
}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\keyword{quantiles}
