% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestData}
\alias{TestData}
\title{Test data}
\format{
A matrix and a vector
}
\usage{
data("TestData")
}
\description{
Test EPIC beta value matrix of 511 CpGs and 10 samples and the sex information.
}
\details{
\itemize{
  \item beta.m: Simulated beta value matrix of 511 CpGs and 10 samples
  \item sex.v: Sex information (0=Female, 1=Male)
}
}
\references{
Bollepalli S, Korhonen T, Kaprio J, Ander S, Ollikainen M.
\emph{EpiSmokEr: a robust classifier to determine smoking status from DNA methylation data.}
Epigenomics (2019) 11(13):1469-1486.
doi:\href{https://doi.org/10.2217/epi-2019-0206}{
10.2217/epi-2019-0206}.
}
\keyword{testdata}
